refactor: reorganize service package structure and update imports
