+++ nostr-java-api/src/main/java/nostr/api/NostrSpringWebSocketClient.java
@@ -3,6 +3,7 @@ package nostr.api;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.NonNull;
+import nostr.api.service.NoteService;
 import nostr.base.IEvent;
 import nostr.base.ISignable;
 import nostr.client.springwebsocket.SpringWebSocketClient;
@@ -12,8 +13,7 @@ import nostr.event.impl.GenericEvent;
 import nostr.event.message.ReqMessage;
 import nostr.id.Identity;
 import nostr.util.NostrUtil;
-import nostr.api.DefaultNoteService;
-import nostr.api.NoteService;
+import nostr.api.service.impl.DefaultNoteService;
 
 import java.io.IOException;
 import java.util.HashMap;
