+++ nostr-java-api/src/main/java/nostr/api/service/NoteService.java
@@ -1,6 +1,7 @@
-package nostr.api;
+package nostr.api.service;
 
 import lombok.NonNull;
+import nostr.api.WebSocketClientHandler;
 import nostr.base.IEvent;
 
 import java.util.List;
